+++ astropy/io/ascii/__init__.py
@@ -21,7 +21,8 @@ from .basic import (Basic,
                     Csv,
                     Tab,
                     NoHeader,
-                    CommentedHeader)
+                    CommentedHeader,
+                    CsvExel)
 from .cds import Cds
 from .latex import Latex, AASTex, latexdicts
 from .html import HTML